[UPDATE] Show following music on personal profile
