+++ public/stylesheets/home.css
@@ -18,10 +18,6 @@ h3,
 h4 {
   width: 100%;
 }
-.user {
-  width: 100px;
-  height: 100px;
-}
 .logo {
   font-style: oblique;
   font-size: 25px;
@@ -109,7 +105,7 @@ h4 {
 .tracks .back:active {
   text-decoration: underline;
 }
-.track-box {
+.tracks .track-box {
   display: flex;
   align-items: center;
   width: 100%;
