+++ routes/index.js
@@ -4,7 +4,5 @@ const controller = require("../controllers/indexController");
 
 router.get("/", controller.getAllCharts);
 router.get("/tracks", controller.getTracks);
-router.get("/follows", controller.getFollows);
-router.post("/follows", controller.setFollows);
 
 module.exports = router;
